Some view changes

+++ Data/Models/Product.cs
@@ -18,10 +18,10 @@
         public string Name { get; set; }
 
         [Required]
-        public double TradePartnerPrice { get; set; }
+        public decimal TradePartnerPrice { get; set; }
 
         [Required]
-        public double Price { get; set; }
+        public decimal Price { get; set; }
 
         public int Quantity { get; set; }
